requirements.txt
aniso8601
hearthstone
hearthstone-data
hslog
MouseInfo
numpy
opencv-python
opencv-contrib-python
Pillow
psutil
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
pywin32
PyYAML
PyQt5
pinyin
keyboard
imutils
python-dotenv
grequests
